requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
ollama>=0.3.0
llama-index-core>=0.11.0
llama-index-llms-ollama>=0.3.0
llama-index-embeddings-ollama>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
html2text>=2024.2.26
tqdm>=4.66.0
